Convolution benchmark is running again.
